+++ src/main/java/seedu/simplst/Capacity.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.simplst;
 
 public enum Capacity {
     SMALL, MEDIUM, LARGE
