+++ src/main/java/seedu/simplst/CommandParsers/StoreCapacityParser.java
