+++ src/main/java/seedu/simplst/LocalStorage.java
@@ -1,8 +1,6 @@
-package seedu.duke;
+package seedu.simplst;
 
-import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.json.simple.JSONValue;
 
 import java.io.FileReader;
 import java.io.FileWriter;
