+++ src/main/java/seedu/simplst/Regex.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.simplst;
 
 import java.util.ArrayList;
 import java.util.HashMap;
