+++ src/test/java/seedu/simplst/CommandsTest.java
@@ -1,8 +1,4 @@
-package seedu.duke;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
+package seedu.simplst;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
