+++ text-ui-test/EXPECTED.TXT
@@ -1,61 +1,4 @@
 Hello from
 Simplst
 What would you like to do?
-New login. Please type the total number of goods your warehouse can hold
-Order 1 is added
-Another command?
-100 chairs are  added
-Another command?
-list command was used wrongly. Type help to see examples
-* Add order `add o/ id/ID r/RECEIVER a/SHIPPING_ADDRESS`
-* Add orderline `add g/ oid/ORDER_ID gid/GOOD_ID n/NAME q/QUANTITY d/DESCRIPTION`
-* Remove order: `remove o/ i/ID`
-* Remove orderline: `remove g/ i/ID q/QUANTITY`
-* List all orders: `list o/`
-* List all goods: `list g/`
-* View order: `view o/ id/ORDER_ID`
-* View orderline: `view g/ id/GOOD_ID`
-* Total quantity of goods: `total`
-Another command?
-Viewing order with id 1
-Receiver: Peter Pan
-Shipping address:NUS
-Items in the order:
-	1. 100 - chairs (this is a chair)
-Another command?
-Viewing item with id 100
-Item name: chairs
-Item description: this is a chair
-Item quantity: 100
-Another command?
-List of orders:
-	1. 1 - Peter Pan (NUS)
-Another command?
-List of goods:
-id | name
-	1. 100 - chairs (this is a chair)
-Another command?
-14 chairs have been removed.
-Another command?
-Viewing order with id 1
-Receiver: Peter Pan
-Shipping address:NUS
-Items in the order:
-	1. 100 - chairs (this is a chair)
-Another command?
-Viewing item with id 100
-Item name: chairs
-Item description: this is a chair
-Item quantity: 86
-Another command?
-There are 86 goods in total.
-Another command?
-Order 1 has been removed.
-Another command?
-There are currently no orders.
-Another command?
-There are no inventory in the warehouse.
-Another command?
-There are 0 goods in total.
-Another command?
 Bye
